Create smart contarcr development
A smart contract is a self-executing contract with the terms of the agreement between buyer and seller being directly written into lines of code. The code and the agreements contained therein are stored and replicated on the Ethereum blockchain.

To write a smart contract on Ethereum, you will need to have a basic understanding of Solidity, the programming language used to write smart contracts on Ethereum.
This simple contract has a public variable called "balance" that is initialized to 100 when the contract is deployed. It has two public functions: "addFunds" and "withdrawFunds". The "addFunds" function allows you to add funds to the contract by incrementing the balance variable, while the "withdrawFunds" function allows you to withdraw funds by decrementing the balance variable. The "withdrawFunds" function also includes a requirement that the amount being withdrawn must not be greater than the current balance, to prevent overdrafts.

To deploy this contract on the Ethereum blockchain, you will need to use a tool like Remix, which is a browser-based Solidity compiler and runtime environment. You can use Remix to write, compile, and deploy your smart contract to the Ethereum network.

I hope this helps! Let me know if you have any further questions. >> https://www.coinsclone.com/